backend/pyproject.toml
[project]
name = "gemini-server"
version = "0.1.0"
description = "FastAPI + GenAI backend"
authors = [
    {name = "BloodCharry",email = "vladvorobei1994@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "google-generativeai (>=0.8.5,<0.9.0)",
    "pillow (>=11.3.0,<12.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
